% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboards.R
\name{getEpochAges}
\alias{getEpochAges}
\title{Get epoch ages for plotting}
\usage{
getEpochAges(
  lakeId,
  chronTable,
  epochFile = "data/Lakes380 DNA and pollen dataset with era - 9 May 2023.xlsx"
)
}
\arguments{
\item{lakeId}{Lakes380 numeric id}

\item{chronTable}{chronTable to update}

\item{epochFile}{epoch file to load}
}
\value{
list of epoch ages
}
\description{
Get epoch ages for plotting
}
