% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSpreadsheet.R
\name{lipd2excel}
\alias{lipd2excel}
\title{Export a lipd object to excel}
\usage{
lipd2excel(L, outname = paste0(L$dataSetName, "-v", getVersion(L), ".xlsx"))
}
\arguments{
\item{L}{a LiPD object}

\item{outname}{name of excel file}
}
\description{
Export a lipd object to excel
}
