% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{mapAgeSummaryToAllPaleoData}
\alias{mapAgeSummaryToAllPaleoData}
\title{Update all the age colums in measurement tables with a summary table}
\usage{
mapAgeSummaryToAllPaleoData(L, ...)
}
\arguments{
\item{L}{a lipd object}

\item{...}{parameters to pass to mapAgeSummaryToPaleoData}
}
\value{
an updated LiPD object
}
\description{
Update all the age colums in measurement tables with a summary table
}
