% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboards.R
\name{plotDashboard}
\alias{plotDashboard}
\title{Plot a Lakes380 data dashboard}
\usage{
plotDashboard(L)
}
\arguments{
\item{L}{a lipd object}
}
\value{
a ggplot object
}
\description{
Plot a Lakes380 data dashboard
}
