% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateLipd.R
\name{updateLipdFromSpreadsheet}
\alias{updateLipdFromSpreadsheet}
\title{Update a LiPD file from a formatted Lakes 380 spreadsheet}
\usage{
updateLipdFromSpreadsheet(L, path)
}
\arguments{
\item{L}{a LiPD object}

\item{path}{path to excel fil}
}
\description{
Update a LiPD file from a formatted Lakes 380 spreadsheet
}
